A lot of updates here There is a problem accessing and applying right sessions in Symfony and I am not sure this is the perfect method but for now it works.
SessionFactory is called everytime a session is needed. It creates a new Session object. The factory is called multiple times , thus session object is new everytime while testing.

To avoid that I wrote a custom session factory that assigns the session to a static variable and returns  everytime the same instance.

The mocked entries are also configured under services.yaml. I have mocked the MockFileSessionStorage which provides a mechanism to store the session in a file.

However, with static session, it may not be needed in future.

Only thing remaining is to reset the session at the end of each test.

Rest of the code is a wip for checkout process