Price getting created using test case
Because I use autocomplete field, and I also use a DTO, during submit the form looks for DTO entity field. I use only keys in DTO wanting to separate the concern as job of DTO is not to validate the keys with database.

So just before submit , we add two form fields and assign them incoming values which are int.

 There is no validation of autocomplete fields as they have been marked as mapped = false

 - Created Fixtures
 - Added SelectElement examples
 - Corrected migration